Newest Assets Updated
I’ve updated the assets again. You now have two sets of assets for the
keyboard. One set has a much larger shadow and that keyboard button
will have a more fluid animation down to the flat level. I have also
uploaded the display ribbons for the code output to go on. There are
two to choose from: one is flat and one has the same shadow as the top
ribbon. Additional mockups for each of the screens have also been
uploaded