Big Changes Everywhere
I changed the original ImageButton to an ImageView. The ImageView has
an onClickListener attached to it that activates an animation sequence
that animates the button. I tried two methods of doing this and settled
on this one. I had a boolean test that would determine whether or not
the image was clicked and then it would change to one particular image,
initially, but I realized that this was extremely ugly and did not look
animated. Also, I redid the layout so that it is now on a
blue_background and added in the assets for the top and bottom output
bars. Also, do not test the app in 1080p right now. Test it on a nexus
4 because I am designing to those specs. The xxhdpi will have a
separate set of drawables and those drawables will need to have
separate parameters. I am considering going back and trying to make the
entire design responsive from the ground up, too.